backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<5.6
django-cors-headers>=4.3,<4.5
psycopg[binary]>=3.1,<3.3
stripe>=10.0,<11.0
gunicorn>=21.2,<23.0
sendgrid